Remove the vendored validation package
The sources and their tests now live in pyvista/input-validation. The
CONTRIBUTING section describing the validation test suites goes with them;
it also referenced tests/core/typing, which no longer exists.
