+++ pyproject.toml
@@ -600,8 +600,6 @@ pyvista = "pv"
 'pyvista/core/_typing_core/_aliases.py' = ['PYI047', 'UP007']
 'pyvista/core/_typing_core/_array_like.py' = ['PYI047', 'UP007']
 'pyvista/core/_typing_core/_dataset_types.py' = ['PYI018']
-'pyvista/core/_validation/_cast_array.py' = ['ANN']
-'pyvista/core/_validation/validate.py' = ['ANN202']
 'pyvista/core/_vtk_utilities.py' = ['ANN']
 'pyvista/core/dataobject.py' = ['PLC0415']
 'pyvista/core/errors.py' = ['ANN']
